add getcpuinfo for ARM, AARCH64, Intel and AMD CPU
differences exist between different architectures in /proc/cpuinfo.
now we move to read /sys/...cpu to figure out the number of cpu, and
read /proc/cpuinfo to find out the type of the CPU.
if we fail to know the CPU type, we return unknown. But anyway, we can
be quiet sure cpunr is always right to support top, mpstat and so on.

Signed-off-by: Barry Song <baohua@linuxep.com>